# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/refresh": {
      "post": {
        "operationId": "refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/logout": {
      "delete": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/job/{id}": {
      "get": {
        "operationId": "find",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        },
        "tags": [
          "job"
        ]
      },
      "put": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        },
        "tags": [
          "job"
        ]
      },
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        },
        "tags": [
          "job"
        ]
      }
    },
    "/api/job": {
      "get": {
        "operationId": "query",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "where",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "count",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobQueryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "job"
        ]
      },
      "put": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        },
        "tags": [
          "job"
        ]
      }
    },
    "/api/company/{id}": {
      "get": {
        "operationId": "find",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          }
        },
        "tags": [
          "company"
        ]
      },
      "put": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          }
        },
        "tags": [
          "company"
        ]
      },
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "company"
        ]
      }
    },
    "/api/company": {
      "get": {
        "operationId": "query",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "where",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "count",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyQueryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "company"
        ]
      },
      "put": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          }
        },
        "tags": [
          "company"
        ]
      }
    },
    "/api/person/{id}": {
      "get": {
        "operationId": "find",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "tags": [
          "person"
        ]
      },
      "put": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePersonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "tags": [
          "person"
        ]
      },
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "tags": [
          "person"
        ]
      }
    },
    "/api/person": {
      "get": {
        "operationId": "query",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "where",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "count",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonQueryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "person"
        ]
      },
      "put": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "tags": [
          "person"
        ]
      }
    }
  },
  "info": {
    "title": "Quasar Nest example",
    "description": "The cats API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "TokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "Job": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "people": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Person"
            }
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "revision": {
            "type": "string"
          }
        },
        "required": [
          "jobId",
          "name",
          "isDeleted",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "revision"
        ]
      },
      "Company": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "people": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Person"
            }
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "revision": {
            "type": "string"
          }
        },
        "required": [
          "companyId",
          "name",
          "isDeleted",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "revision"
        ]
      },
      "Person": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "jobId": {
            "type": "string"
          },
          "job": {
            "$ref": "#/components/schemas/Job"
          },
          "companyId": {
            "type": "string"
          },
          "company": {
            "$ref": "#/components/schemas/Company"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "revision": {
            "type": "string"
          }
        },
        "required": [
          "personId",
          "avatar",
          "firstName",
          "lastName",
          "email",
          "jobId",
          "companyId",
          "isDeleted",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "revision"
        ]
      },
      "JobQueryResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Job"
            }
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "data"
        ]
      },
      "ConnectPersonDto": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "string"
          }
        },
        "required": [
          "personId"
        ]
      },
      "CreateJobPeopleRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectPersonDto"
            }
          }
        },
        "required": [
          "connect"
        ]
      },
      "CreateJobDto": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "people": {
            "$ref": "#/components/schemas/CreateJobPeopleRelationInputDto"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateJobPeopleRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectPersonDto"
            }
          }
        },
        "required": [
          "connect"
        ]
      },
      "UpdateJobDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "people": {
            "$ref": "#/components/schemas/UpdateJobPeopleRelationInputDto"
          }
        }
      },
      "CompanyQueryResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Company"
            }
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "data"
        ]
      },
      "CreateCompanyPeopleRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectPersonDto"
            }
          }
        },
        "required": [
          "connect"
        ]
      },
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "people": {
            "$ref": "#/components/schemas/CreateCompanyPeopleRelationInputDto"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCompanyPeopleRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectPersonDto"
            }
          }
        },
        "required": [
          "connect"
        ]
      },
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "people": {
            "$ref": "#/components/schemas/UpdateCompanyPeopleRelationInputDto"
          }
        }
      },
      "PersonQueryResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Person"
            }
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "data"
        ]
      },
      "ConnectJobDto": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          }
        },
        "required": [
          "jobId"
        ]
      },
      "CreatePersonJobRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "$ref": "#/components/schemas/ConnectJobDto"
          }
        },
        "required": [
          "connect"
        ]
      },
      "ConnectCompanyDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          }
        },
        "required": [
          "companyId"
        ]
      },
      "CreatePersonCompanyRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "$ref": "#/components/schemas/ConnectCompanyDto"
          }
        },
        "required": [
          "connect"
        ]
      },
      "CreatePersonDto": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "job": {
            "$ref": "#/components/schemas/CreatePersonJobRelationInputDto"
          },
          "company": {
            "$ref": "#/components/schemas/CreatePersonCompanyRelationInputDto"
          }
        },
        "required": [
          "avatar",
          "firstName",
          "lastName",
          "email",
          "job",
          "company"
        ]
      },
      "UpdatePersonJobRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "$ref": "#/components/schemas/ConnectJobDto"
          }
        },
        "required": [
          "connect"
        ]
      },
      "UpdatePersonCompanyRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "$ref": "#/components/schemas/ConnectCompanyDto"
          }
        },
        "required": [
          "connect"
        ]
      },
      "UpdatePersonDto": {
        "type": "object",
        "properties": {
          "avatar": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "job": {
            "$ref": "#/components/schemas/UpdatePersonJobRelationInputDto"
          },
          "company": {
            "$ref": "#/components/schemas/UpdatePersonCompanyRelationInputDto"
          }
        }
      }
    }
  }
}